# backend/docs/api/openapi.pretty.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Niyat Ride API",
        "description": "API Documentation for Niyat Ride",
        "contact": {
            "name": "Niyat Support",
            "email": "backendteam@gmail.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
...
        "/api/admin/pricing/base-price": {
            "get": {
                "tags": [
                    "Base Price Management"
                ],
                "summary": "Get current base price",
                "description": "Retrieve the current active base price with details",
                "operationId": "getCurrentBasePrice",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/BasePriceResponseDTO"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Base Price Management"
                ],
                "summary": "Update base price",
                "description": "Update the base price for all rides",
                "operationId": "updateBasePrice",
                "parameters": [
                    {
                        "name": "adminId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BasePriceUpdateDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/BasePriceResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/drivers/signup/verify-otp": {
            "post": {
                "tags": [
                    "Driver Management"
                ],
                "summary": "Verify OTP and complete driver signup",
                "operationId": "verifySignupOtp",
                "parameters": [
                    {
                        "name": "phoneNumber",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "otp",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/DriverResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/drivers/signup/request-otp": {
            "post": {
                "tags": [
                    "Driver Management"
                ],
                "summary": "Request OTP for driver signup",
                "operationId": "requestSignupOtp",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DriverSignupDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/drivers/login/verify-otp": {
            "post": {
                "tags": [
                    "Driver Management"
                ],
                "summary": "Verify OTP and log in driver",
                "operationId": "verifyLoginOtp",
                "parameters": [
                    {
                        "name": "phoneNumber",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "otp",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/DriverResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/drivers/login/request-otp": {
            "post": {
                "tags": [
                    "Driver Management"
                ],
                "summary": "Request OTP for driver login",
                "operationId": "requestLoginOtp",
                "parameters": [
                    {
                        "name": "phoneNumber",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/customers/signup/verify-otp": {
            "post": {
                "tags": [
                    "Customer Management"
                ],
                "summary": "Verify OTP and complete customer signup",
                "operationId": "verifySignupOtp_1",
                "parameters": [
                    {
                        "name": "phoneNumber",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "otp",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomerResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/customers/signup/request-otp": {
            "post": {
                "tags": [
                    "Customer Management"
                ],
                "summary": "Request OTP for customer signup",
                "operationId": "requestSignupOtp_1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CustomerSignupDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/customers/login/verify-otp": {
            "post": {
                "tags": [
                    "Customer Management"
                ],
                "summary": "Verify OTP and log in customer",
                "operationId": "verifyLoginOtp_1",
                "parameters": [
                    {
                        "name": "phoneNumber",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "otp",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomerResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/customers/login/request-otp": {
            "post": {
                "tags": [
                    "Customer Management"
                ],
                "summary": "Request OTP for customer login",
                "operationId": "requestLoginOtp_1",
                "parameters": [
                    {
                        "name": "phoneNumber",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/customers/updateCustomer/{customerId}": {
            "patch": {
                "tags": [
                    "Customer Management"
                ],
                "summary": "Update customer details",
                "operationId": "updateCustomer",
                "parameters": [
                    {
                        "name": "customerId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CustomerUpdateDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomerResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BasePriceUpdateDTO": {
                "required": [
                    "amount"
                ],
                "type": "object",
                "properties": {
                    "amount": {
                        "minimum": 0.01,
                        "exclusiveMinimum": false,
                        "type": "number"
                    },
                    "changeReason": {
                        "maxLength": 500,
                        "minLength": 0,
                        "type": "string"
                    }
                }
            },
            "BasePriceResponseDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "amount": {
                        "type": "number"
                    },
                    "currency": {
                        "type": "string"
                    },
                    "isActive": {
                        "type": "boolean"
                    },
                    "changeReason": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedByAdminName": {
                        "type": "string"
                    }
                }
            },
...
            "DriverResponseDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phoneNumber": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "DISPATCHER",
                            "DRIVER",
                            "Admin",
                            "CUSTOMER"
                        ]
                    },
                    "status": {
                        "type": "string"
                    },
                    "licenseNumber": {
                        "type": "string"
                    },
                    "vehicleType": {
                        "type": "string"
                    },
                    "vehiclePlateNumber": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "DriverSignupDTO": {
                "required": [
                    "backLicenceImage",
                    "email",
                    "firstName",
                    "frontLicenceImage",
                    "lastName",
                    "licenseNumber",
                    "licesneExipirationDate",
                    "phoneNumber",
                    "vehicleModel",
                    "vehiclePlateNumber"
                ],
                "type": "object",
                "properties": {
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phoneNumber": {
                        "pattern": "^[0-9]{9,15}$",
                        "type": "string"
                    },
                    "licenseNumber": {
                        "type": "string"
                    },
                    "vehicleModel": {
                        "type": "string"
                    },
                    "vehiclePlateNumber": {
                        "type": "string"
                    },
                    "frontLicenceImage": {
                        "type": "string"
                    },
                    "backLicenceImage": {
                        "type": "string"
                    },
                    "licesneExipirationDate": {
                        "type": "string",
                        "format": "date"
                    }
                }
            },
...
            "CustomerResponseDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phoneNumber": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "DISPATCHER",
                            "DRIVER",
                            "Admin",
                            "CUSTOMER"
                        ]
                    },
                    "status": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "CustomerSignupDTO": {
                "required": [
                    "phoneNumber"
                ],
                "type": "object",
                "properties": {
                    "phoneNumber": {
                        "type": "string"
                    }
                }
            },
...
            "CustomerUpdateDTO": {
                "required": [
                    "firstName",
                    "lastName"
                ],
                "type": "object",
                "properties": {
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}